feat(pg_stage): Refactoring
Move constants, dataclasses and exceptions in files
