+++ src/pg_stage/exceptions.py
@@ -1,0 +1,2 @@
+class PgDumpError(Exception):
+    """Базовое исключение для ошибок обработки дампов PostgreSQL."""
